pyproject.toml
[project]
name = "hf-model-downloader-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "huggingface-hub>=0.28.1",
    "python-dotenv>=1.0.1",
]
